Optimize for check whether a path is a rect
Checking whether a path is very expensive.  It goes through each verb in
the pathref, and compute the contour.

SkPath::isRect() is called in each SkGpuDevice::drawPath() to check whether
we can route drawPath() to drawRect().  But SkPath::isRect() is too
expensive.

We create an internal flag fPathType.  fPathType is set to rect_type if
(1) SkPath::addRect() is called once

fPathType is set to possible quad if
(1) SkPath::lineTo() called 3 times and SkPath::close() is called
(2) SkPath::lineTo() called 4 times (I know this is not a precise).

In SkGpuDevice::drawPath(), we first call SkPath::isQuad() to filter
out those paths that are not possibly to be rect. Ony those paths that
are quad, then we call SkPath::isRect()

We also optimize SkPath::isRect().  If fPathType is rect_type, we can
get rect quickly and return true